+++ gno.land/pkg/keyscli/maketx.go
@@ -22,12 +22,12 @@ func NewMakeTxCmd(rootCfg *client.BaseCfg, io commands.IO) *commands.Command {
 
 	cmd.AddSubCommands(
 		client.NewMakeSendCmd(cfg, io),
+		client.NewSessionCmd(cfg, io),
 
 		// custom commands
 		NewMakeAddPkgCmd(cfg, io),
 		NewMakeCallCmd(cfg, io),
 		NewMakeRunCmd(cfg, io),
-		NewSessionCmd(cfg, io),
 	)
 
 	return cmd
