+++ tm2/pkg/crypto/keys/client/maketx.go
@@ -63,6 +63,7 @@ func NewMakeTxCmd(rootCfg *BaseCfg, io commands.IO) *commands.Command {
 
 	cmd.AddSubCommands(
 		NewMakeSendCmd(cfg, io),
+		NewSessionCmd(cfg, io),
 	)
 
 	return cmd
