+++ tm2/pkg/crypto/keys/client/session_revokeall.go
@@ -1,4 +1,4 @@
-package keyscli
+package client
 
 import (
 	"context"
@@ -7,7 +7,6 @@ import (
 	"github.com/gnolang/gno/tm2/pkg/amino"
 	"github.com/gnolang/gno/tm2/pkg/commands"
 	"github.com/gnolang/gno/tm2/pkg/crypto/keys"
-	"github.com/gnolang/gno/tm2/pkg/crypto/keys/client"
 	"github.com/gnolang/gno/tm2/pkg/errors"
 	"github.com/gnolang/gno/tm2/pkg/sdk/auth"
 	"github.com/gnolang/gno/tm2/pkg/std"
@@ -82,7 +81,7 @@ func execSessionRevokeAll(cfg *SessionRevokeAllCfg, args []string, io commands.I
 	}
 
 	if cfg.RootCfg.RootCfg.Broadcast {
-		err := client.ExecSignAndBroadcast(cfg.RootCfg.RootCfg, args, tx, io)
+		err := ExecSignAndBroadcast(cfg.RootCfg.RootCfg, args, tx, io)
 		if err != nil {
 			return err
 		}
